reset password update | using code instead of link for reset password

+++ app/Repositories/UserRepository.php
@@ -39,17 +39,22 @@ abstract class UserRepository {
         return Socialite::driver($driver)->stateless()->user()->user;
     }
 
-    public function resetPassword($new_password, String $token, Model &$user) {
-        $record = PasswordReset::where('email', $user->email)->where('token', $token)->first();
+    public function resetPassword($new_password, String $token, Model &$user, $user_type) {
+        $record = PasswordReset::where('email', $user->email)
+            ->where('user_type', $user_type)->first();
 
         if(!$record) {
-            return false;
+            return -1;
+        }
+
+        if($record->token != $token) {
+            return -2;
         }
 
         $record->delete();
 
         $user->password = bcrypt($new_password);
 
-        return true;
+        return 1;
     }
 }
